Refactor Controller and Services
